+++ tests/test_rewrite.py
@@ -1,6 +1,6 @@
 import ast
 
-from tests.test_suite import testable_test
+from tests.utilities import testable_test
 from ward import test, fixture
 from ward.rewrite import (
     rewrite_assertions_in_tests,
